Fix a harmless compiler warning in the sessionAppendDelete() function.
FossilOrigin-Name: 1324d08d0c639378cf1b03bd6410db81947747ac

+++ ext/session/sqlite3session.c
@@ -1621,7 +1621,7 @@ static int sessionAppendDelete(
         }
       }
       if( abPK[i] ){
-        sessionAppendBlob(pBuf, pStart, a-pStart, &rc);
+        sessionAppendBlob(pBuf, pStart, (int)(a-pStart), &rc);
       }
     }
     assert( (a - p->aRecord)==p->nRecord );
